storm_webapp/requirements_webapp.txt
Django>=5.0,<6.0
channels>=4.0,<5.0
daphne>=4.0,<5.0
djangorestframework>=3.15,<4.0
pywhatkit>=5.4
python-decouple>=3.8
